Change approach to venv
The point of this commit was to create a virtualenv with a known
name. When relying on poetry previously, it creates a slug name
but when you want to copy that venv to the production image
(that doesn't have poetry) it is hard to know what that folder
name was.

Poetry installs straight into a virtualenv if that virtualenv
is already activated, therefore this approach creates the venv
with a known name and activates it prior to calling poetry
commands.

The POETRY_VIRTUALENVS_PATH is maintained because it seems to
help VSCode suggest the venv created even though it was not
poetry that created the initial venv