Remove the virtual env activation problem
Activating a virtual environment is mainly a "convenience". The
only thing a docker image really needs to be concerned with is
having the venv at the front of path, so that python is executed
from there. So instead of trying to activate environment for
commands, this now makes sure the PATH is setup with the venv
location.

This leans upon poetry's behavior of checking VIRTUAL_ENV
to find out whether a virtual environment is activated or not.
By setting VIRUTAL_ENV, poetry makes the assumption it is and
installs straight to there instead of creating a new environment.

reference that helped greatly:
https://pythonspeed.com/articles/activate-virtualenv-dockerfile/